Update dependencies, refactor styles, and clean up imports
- Upgraded Next.js to version 15.1.4 and updated related dependencies in package.json.
- Refactored global styles by removing commented-out CSS rules.
- Enhanced layout styles for better visibility and user experience.
- Adjusted import paths in action.js and page.tsx to align with the new project structure.
