+++ app/globals.css
@@ -15,8 +15,3 @@
     --background-end-rgb: 0, 0, 0;
   }
 }
-/* 
-body {
-  color: rgb(var(--foreground-rgb));
-
-} */
