+++ app/layout.tsx
@@ -73,12 +73,10 @@ export default function RootLayout({
             {/* Content */}
             {children}
             {/* Footer */}
-            <div className="text-black bg-gray-300 w-full p-4 text-sm grid place-items-center mt-auto">
+            <div className="text-white bg-slate-900 w-full p-8 text-sm grid place-items-center">
               <p>
                 &copy; Animeflix {new Date().getFullYear()} by&nbsp;
-                <a href="https://github.com/Sodiumchloridy">
-                  <span className="text-blue-500">Sodiumchloridy</span>
-                </a>
+                <a className="text-blue-500 hover:text-blue-600" href="https://github.com/Sodiumchloridy">Sodiumchloridy</a>
                 .
               </p>
               <p>
